+++ brainbox/behavior/__init__.py
@@ -1,2 +1,1 @@
-from .behavior import *
 import wheel