+++ examples/one/histology/list_trajectory_more1alignment.py
@@ -1,0 +1,31 @@
+'''
+List sessions with 2 or more histology-ephys alignment done.
+TODO: remove session with multiple alingment done by 1 user only.
+'''
+# Author: Gaelle Chapuis
+from oneibl.one import ONE
+import numpy as np
+import pandas as pd
+one = ONE()
+rec_with_hist = one.alyx.rest('trajectories', 'list', provenance='Ephys aligned histology track')
+eids = np.array([s['id'] for s in rec_with_hist])
+
+json = [s['json'] for s in rec_with_hist]
+idx_none = [i for i, val in enumerate(json) if val is None]
+json_val = np.delete(json, idx_none)
+keys = [list(s.keys()) for s in json_val]
+
+# user_key = [s[0][20:] for s in keys]  # remove date-time todo
+
+# Find index of json fields with 2 or more keys
+len_key = [len(s) for s in keys]
+idx_several = [i for i, val in enumerate(len_key) if val >= 2]
+eid_several = eids[idx_several]
+
+# create dataframe
+frame = pd.DataFrame()
+frame['eid'] = eid_several
+# frame['user'] = np.array(user_key)[idx_several]
+frame['key'] = np.array(keys)[idx_several]
+
+print(f'{frame}')